Fix delayed cursor default moving from search line (#45)
There was previously a delay in defaulting the cursor position to the
end of the line when moving from the search line, to the line below, if
the line below had a hidden prefix. This sets the lenHiddenMatchPrefix
variable a second time in the loop which is required for the cursor
position resolution.