Fix version tagging for build from source (#77)
At release time, a build tag is now generated and flushed to a `VERSION`
file which is written to the project root. In the Makefile, when
attempting to generate the tag, it will initially attempt to build from
the git log, but if this fails (which it will if `make build` is being
run from the packaged release source code), it will instead look to this
file to retrieve the tag. This isn't fool-proof, but works if git is
maintained as the source of truth for tagging (which is the intention)