Simplify non-git build tagging (#78)
* Simplify non-git build tagging

Use current time when generating the build tag when building from source
outside of a git-repo. The tag is still inferred from the `VERSION`
file, however.

It got confusing trying to coordinate build tags inferred from git
commits with the `VERSION` file in the repo. In short, it's impossible
to enter the commit dt into a file, and have that file within the scope
of that commit (or at least my git-fu isn't good enough to know how to
wrangle it). So I've given up for now and will settle with this
solution.

* Add explicit Makefile steps to allow consistent git tag and VERSION file contents